Intro and Gem Count Working
Coded the Intro scene and the GEM count. All other scenes are stand ins that have the gem count coded to carry over from the intro scene. I still need to reset it if the game is played multiple times. Advance buttons can be replaced by a decorative button. It is a prefab, so it will affect all instances when changed. The GEM count image can be replaced by a 2D sprite of a gem later. Start screen also has stand in text and buttons.