Added Snake and Traveler to Backgrounds. Added New Font to Titles
Added snake backgrounds to End 1 and End 5. Added Traveler to the desert scene at the beginning. Added new font to the game title, credits title, and "the end" on all the endings. Also listed this font in the credits. Fixed a few places where the incorrect backgrounds were turned on for some reason.